Split routing methods
